requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# classifier.py: 4
# data_preprocessing.py: 7,8
Keras == 2.1.1

# sniffer.py: 8
NetfilterQueue == 0.8.1

# data_preprocessing.py: 9
h5py == 2.7.1

# classifier.py: 3
# data_preprocessing.py: 3
numpy == 1.13.3

# data_preprocessing.py: 6
pandas == 0.21.0

# iptables.py: 1
python_iptables == 0.12.0

# scapy3-sniffer.py: 2
# sniffer.py: 6
scapy_http == 1.8

# scapy3-sniffer.py: 1
# scapy_bridge.py: 9
# scapy_test.py: 1
# sniffer.py: 5
# tcp_http_spoof.py: 1
scapy_python3 == 0.21

# sniffer.py: 7
scapy_ssl_tls == 1.3.0

# data_preprocessing.py: 1
scikit_learn == 0.19.1
